refactor(epoch-sync): Move EpochSync responses to a separate file (#9593)
